Move root crate into maplibre-core
