+++ maplibre-core/src/error.rs
