+++ maplibre-core/src/input/pan_handler.rs
