+++ maplibre-core/src/input/pinch_handler.rs
