+++ maplibre-core/src/input/query_handler.rs
