+++ maplibre-core/src/input/shift_handler.rs
