+++ maplibre-core/src/input/tilt_handler.rs
