+++ maplibre-core/src/input/zoom_handler.rs
