+++ maplibre-core/src/io/geometry_index.rs
