+++ maplibre-core/src/io/scheduler.rs
