+++ maplibre-core/src/io/shared_thread_state.rs
