+++ maplibre-core/src/io/static_tile_fetcher.rs
