+++ maplibre-core/src/io/tile_cache.rs
