+++ maplibre-core/src/platform/noweb/schedule_method.rs
