+++ maplibre-core/src/platform/web/http_client.rs
