+++ maplibre-core/src/platform/web/legacy_webworker_fetcher.rs
