+++ maplibre-core/src/platform/web/pool.rs
