+++ maplibre-core/src/platform/web/schedule_method.rs
