+++ maplibre-core/src/render/buffer_pool.rs
