+++ maplibre-core/src/render/shaders/tile.fragment.wgsl
