+++ maplibre-core/src/render/shaders/tile_mask.fragment.wgsl
