+++ maplibre-core/src/render/shaders/tile_mask.vertex.wgsl
