+++ maplibre-core/src/render/texture.rs
