+++ maplibre-core/src/render/tile_view_pattern.rs
