+++ maplibre-core/src/tessellation/zero_tessellator.rs
