+++ maplibre-core/src/util/fps_meter.rs
