+++ maplibre-core/src/util/grid.rs
