+++ maplibre-core/src/util/math.rs
