+++ maplibre-core/src/window.rs
